@@ -1,5 +1,9 @@
-module.exports = {
-  ACLService: require('./src/ACLService'),
-  ACLError: require('./src/ACLError'),
-  ACLManager: require('./src/ACLManager')
+import ACLService from './src/ACLService.js'
+import ACLError from './src/ACLError.js'
+import ACLManager from './src/ACLManager.js'
+
+export {
+  ACLService,
+  ACLError,
+  ACLManager
 }
